Organization of methods
Moved extension methods into separate package
