+++ .gitignore
@@ -3,3 +3,5 @@ config/*.yaml
 .idea/
 
 \.DS_Store
+
+out/
