+++ src/test/kotlin/com/immanuelqrw/gitbase/configuration/ConfigurationTest.kt
@@ -1,0 +1,4 @@
+package com.immanuelqrw.gitbase.configuration
+
+
+class ConfigurationTest
